refactor(vue): group markdown renderers
