+++ packages/vue/src/components/renderers/markdown/node-styles.ts
